Repair entity placement
The entity placement wrongly depended on the paintsize.
This was fine when it was always 1, but now that it is variable
the render location was wrong.

+++ Entities.h
@@ -12,8 +12,8 @@
 struct EntityProperties {
     float size;
     short textureId;
-    int tileTop;
-    int tileLeft;
+    short tileTop;
+    short tileLeft;
     float distanceFromCeiling;
     short type;
 };
